setup connection pool and build routers
