+++ static/css_js_folder/css/attraction_style.css
