+++ static/css_js_folder/js/attraction_script.js
