+++ static/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="/static/css_js_folder/index_style.css" />
+    <link rel="stylesheet" href="/static/css_js_folder/css/index_style.css" />
     <title>Index</title>
   </head>
   <body>
@@ -85,7 +85,7 @@
     <footer class="foot">
       <p class="foot-p">COPYRIGHT © 2021 台北一日遊</p>
     </footer>
-    <script src="/static/css_js_folder/index_script.js"></script>
+    <script src="/static/css_js_folder/js/index_script.js"></script>
 
     </body>
 </html>
